Updated templates for base URL handling
This replaces all hard-coded or hand-assembled URLs in the templates
with invocations of url_for. The url_for interface is quite simple;
usually, you can just wrap the string URL (in "absolute path" form,
i.e. beginning with /) in c.url_for() and it will work.

This is meant to cover all URLs emitted by the backend in template
code. The next thing to cover is URLs in the frontend JS code.