Updated frontend code for base URL handling
This is by the mechanism of a new class, apiURL, defined in the
common.js file. The apiURL class reads the URL prefix from the backend
and prepends it onto any supplied URL at time of use. Any URL pointing
into the app should be generated through apiURL in order to respect
the URL prefix.