Spacebar keybind improvements - Snapping for all views & Fix broken next page function
The new spacebar logic to go to next page doesn't work on Firefox at all, whereas the previous logic only worked on some pages, as I theorized it was caused by fractional pixels, images that rounded up or down to a different pixel count caused a mismatch on where LRR thought the end of the page was.

Fixing that could've been as simple as
```
if (($(window).height() + $(window).scrollTop()) >= (LRR.getDocHeight() - 3)) {
    (Reader.mangaMode) ? Reader.changePage(-1) : Reader.changePage(1);
}
```
But I decided to make the spacebar logic better, now you can easily spam it to keep going down, snap to bottom of images, or skip to next image if already at the bottom. And the opposite for going up and to previous pages with shift+space.